replace randomHelper
use new version randomHelper and randomExtension.

+++ Common/CommonDotNet/System/RandomHelper/RandomHelper.NextSingle.cs
@@ -6,31 +6,31 @@ namespace System
         /// <summary>
         /// 返回非负随机数。
         /// </summary>
-        /// <returns>大于等于零且小于 System.Double.MaxValue 的双精度浮点数。</returns>
-        public static double NextDouble()
+        /// <returns>大于等于零且小于 System.Single.MaxValue 的单精度浮点数。</returns>
+        public static float NextSingle()
         {
-            return Rand.NextDouble();
+            return Rand.NextSingle();
         }
 
         /// <summary>
         /// 返回随机数。
         /// </summary>
         /// <param name="containNegative">是否包含负数。</param>
-        /// <returns>返回一个随机的双精度浮点数。</returns>
-        public static double NextDouble(bool containNegative)
+        /// <returns>返回一个随机的单精度浮点数。</returns>
+        public static float NextSingle(bool containNegative)
         {
-            return Rand.NextDouble(containNegative);
+            return Rand.NextSingle(containNegative);
         }
 
         /// <summary>
         /// 返回一个小于所指定最大值的非负随机数。
         /// </summary>
         /// <param name="maxValue">要生成的随机数的上限（随机数不能取该上限值）。maxValue 必须大于或等于零。</param>
-        /// <returns>大于等于零且小于 maxValue 的双精度浮点数，即：返回值的范围通常包括零但不包括 maxValue。不过，如果 maxValue 等于零，则返回 maxValue。</returns>
+        /// <returns>大于等于零且小于 maxValue 的单精度浮点数，即：返回值的范围通常包括零但不包括 maxValue。不过，如果 maxValue 等于零，则返回 maxValue。</returns>
         /// <exception cref="System.ArgumentOutOfRangeException"><c>maxValue</c> 小于 0。</exception>
-        public static double NextDouble(double maxValue)
+        public static float NextSingle(float maxValue)
         {
-            return Rand.NextDouble(maxValue);
+            return Rand.NextSingle(maxValue);
         }
 
         /// <summary>
@@ -38,11 +38,11 @@ namespace System
         /// </summary>
         /// <param name="minValue">返回的随机数的下界（随机数可取该下界值）。</param>
         /// <param name="maxValue">返回的随机数的上界（随机数不能取该上界值）。maxValue 必须大于或等于 minValue。</param>
-        /// <returns>一个大于等于 minValue 且小于 maxValue 的双精度浮点数，即：返回的值范围包括 minValue 但不包括 maxValue。如果 minValue 等于 maxValue，则返回 minValue。</returns>
+        /// <returns>一个大于等于 minValue 且小于 maxValue 的单精度浮点数，即：返回的值范围包括 minValue 但不包括 maxValue。如果 minValue 等于 maxValue，则返回 minValue。</returns>
         /// <exception cref="System.ArgumentOutOfRangeException"><c>minValue</c> 大于 <c>maxValue</c>。</exception>
-        public static double NextDouble(double minValue, double maxValue)
+        public static float NextSingle(float minValue, float maxValue)
         {
-            return Rand.NextDouble(minValue, maxValue);
+            return Rand.NextSingle(minValue, maxValue);
         }
     }
 }
